tensormap-server/pyproject.toml
[tool.poetry]
name = "tensormap-server"
version = "0.1.0"
description = "TensorMap Python Server"
authors = ["Saatwik Vasishtha <saatwik.vasishtha@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
alembic = "1.6.2"
python = "^3.9"
numpy = "^1.19.5"
pandas = "1.2.4"
flask = "2.0.3"
flask-cors = "3.0.10"
flask-restful = "0.3.9"
flask-migrate = "3.0.0"
flask-sqlalchemy = "2.5.1"
flask-socketio = "5.1.1"
tensorflow = "2.5.0"
tensorboard = "2.5.0"
python-dotenv = "0.17.1"
pyyaml = "5.4.1"
flatten-json = "0.1.13"
python-decouple = "3.8"
psycopg2-binary = "^2.9.9"
Werkzeug = "2.0.3"
protobuf = "3.17.3"
rsa = "4.7.2"
zope-event = "4.6"
python-socketio = "5.4.1"
python-engineio = "4.2.1"
pytest = "^8.3.4"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.6"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"

[tool.ruff]
select = ["E", "F", "W", "I", "B", "C", "N", "UP", "YTT"]  # Choose your rules
ignore = []  # Add any rules you want to ignore
line-length = 88  # Set your desired line length

[tool.ruff.isort]
known-first-party = ["tensormap_server"]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
